requirements.txt
python-dotenv==1.0.1
streamlit==1.30.0
google-generativeai==0.3.2